Move the RecommendationChecker to its own file
